+++ packages/__tests__/src/view-render/timegrid-slots.ts
@@ -1,0 +1,9 @@
+describe('timegrid slots', () => {
+  // https://github.com/fullcalendar/fullcalendar/issues/5952
+  it('can render a single big slot without error', () => {
+    initCalendar({
+      initialView: 'timeGridDay',
+      slotDuration: '24:00',
+    })
+  })
+})